Create LICENSE
I like your mod very much! I wish you good development! But I would like it to be under a free license for people! This will allow to create content for it without fear of being blocked! (like you did for Create, which is also licensed by MIT).
As a compromise, I suggest you keep the copyright only on the mod name and its logo, and keep the code under MIT 
Thank you!